Move create_connection_cache utility function to a more appropriate location
