+++ tps-client/src/lib.rs
@@ -148,3 +148,4 @@ pub trait TpsClient {
 mod bank_client;
 mod rpc_client;
 mod tpu_client;
+pub mod utils;
